requirements.txt
google-generativeai
PyMuPDF
reportlab